修正合約存檔里程數問題

+++ Domain/SP/BE/Input/SPInput_BE_OrderModify.cs
@@ -36,11 +36,11 @@ namespace Domain.SP.BE.Input
         /// <summary>
         /// 開始里程
         /// </summary>
-        public int start_mile { set; get; }
+        public float start_mile { set; get; }
         /// <summary>
         /// 結束里程
         /// </summary>
-        public int end_mile { set; get; }
+        public float end_mile { set; get; }
         /// <summary>
         /// 逾時費
         /// </summary>
